fix: edit attribute and method point type

+++ src/lib/Valuable/Player/Player.cpp
@@ -49,7 +49,7 @@ string Player::getNickname() const
     return this->nickname_;
 }
 
-int Player::getPoint() const
+long long Player::getPoint() const
 {
     return this->point_;
 }
